api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.5.2
jinja2==3.1.2
pdfkit==1.0.0
openai==1.6.1
python-multipart==0.0.9
yt-dlp==2024.4.9
webvtt-py==0.4.6
